+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/table/ResolvedTableBuilderInternal.java
@@ -14,24 +14,11 @@
  * limitations under the License.
  */
 
-package io.delta.kernel.internal;
+package io.delta.kernel.internal.table;
 
-import io.delta.kernel.ResolvedTable;
-import io.delta.kernel.internal.actions.Metadata;
-import io.delta.kernel.internal.actions.Protocol;
-import io.delta.kernel.internal.annotation.VisibleForTesting;
-import io.delta.kernel.internal.snapshot.LogSegment;
+import io.delta.kernel.ResolvedTableBuilder;
 import io.delta.kernel.internal.util.Clock;
 
-public interface ResolvedTableInternal extends ResolvedTable {
-  String getLogPath();
-
-  Protocol getProtocol();
-
-  Metadata getMetadata();
-
-  Clock getClock();
-
-  @VisibleForTesting
-  LogSegment getLogSegment();
+public interface ResolvedTableBuilderInternal extends ResolvedTableBuilder {
+  ResolvedTableBuilderInternal withClock(Clock clock);
 }
